// chpqa-app/src/app/registration/registration-confirmation/models/registration-confirmation-path.model.ts
export enum RegistrationConfirmationPath {
  BASE_PATH = '/registration-confirmation',
  EMAIL_CONFIRMATION_SUCCESS = '',
  ENTER_RESPONSIBLE_PERSON = 'enter-details',
  CHOOSE_PASSWORD = 'choose-password',
  CHECK_ANSWERS = 'check-answers',
  SUCCESS = 'success',
  CONFIRM_ORGANISATION_ADDRESS = 'confirm-organisation-address',
  SEARCH_ORGANISATION_ADDRESS = 'search-organisation-address',
  SELECT_ORGANISATION_ADDRESS = 'select-organisation-address',
  ADD_ORGANISATION_ADDRESS = 'add-organisation-address',
  LEGAL_STATEMENT = 'legal-statement',
  TERMS_CONDITIONS='terms-conditions'
}